Move Color.toAwtColor into the ui-graphics module
