+++ compose/ui/ui-graphics/src/desktopMain/kotlin/androidx/compose/ui/graphics/awt/AwtColor.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.graphics.awt
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.graphics.Color
+
+/**
+ * Converts a Compose Color to its corresponding [java.awt.Color].
+ *
+ * Note that the actual color drawn on the screen by Compose and AWT may differ slightly due to
+ * differences in color space.
+ */
+@ExperimentalComposeUiApi
+fun Color.toAwtColor() = java.awt.Color(red, green, blue, alpha)
